Improved notification system #351
In order to improve the notitication messages and provide some additional data, the notification system had to be improved. Now a :data field may be provides for the notify_all_users method to provide up to 2 additional strings with data to display in the notification message.

Additionally there's a better notification message for name changes now.